+++ glean-core/ios/GleanTests/GleanTests.swift
@@ -261,7 +261,8 @@ class GleanTests: XCTestCase {
             includeInfoSections: true,
             enabled: true,
             schedulesPings: [],
-            reasonCodes: []
+            reasonCodes: [],
+            followsCollectionEnabled: true
         )
 
         let counter = CounterMetricType(CommonMetricData(
@@ -283,7 +284,7 @@ class GleanTests: XCTestCase {
         Glean.shared.resetGlean(clearStores: false)
 
         // Glean might still be initializing. Disable upload.
-        Glean.shared.setUploadEnabled(false)
+        Glean.shared.setCollectionEnabled(false)
 
         // Set data and try to submit a custom ping.
         counter.add(1)
@@ -401,7 +402,8 @@ class GleanTests: XCTestCase {
             includeInfoSections: true,
             enabled: true,
             schedulesPings: [],
-            reasonCodes: []
+            reasonCodes: [],
+            followsCollectionEnabled: true
         )
 
         let counter = CounterMetricType(CommonMetricData(