+++ glean-core/rlb/examples/delayed-ping-data.rs
@@ -74,8 +74,19 @@ impl net::PingUploader for MovingUploader {
 }
 
 #[allow(non_upper_case_globals)]
-pub static PrototypePing: Lazy<PingType> =
-    Lazy::new(|| PingType::new("prototype", true, true, false, true, true, vec![], vec![]));
+pub static PrototypePing: Lazy<PingType> = Lazy::new(|| {
+    PingType::new(
+        "prototype",
+        true,
+        true,
+        false,
+        true,
+        true,
+        vec![],
+        vec![],
+        true,
+    )
+});
 
 fn main() {
     env_logger::init();
@@ -85,6 +96,7 @@ fn main() {
     let data_path = PathBuf::from(args.next().expect("need data path"));
     let state = args.next().unwrap_or_default();
 
+    _ = &*PrototypePing;
     let uploader = MovingUploader::new(data_path.clone());
     let cfg = ConfigurationBuilder::new(true, data_path, "glean.pingflush")
         .with_server_endpoint("invalid-test-host")