+++ glean-core/rlb/examples/long-running.rs
@@ -40,8 +40,19 @@ impl net::PingUploader for FakeUploader {
 }
 
 #[allow(non_upper_case_globals)]
-pub static PrototypePing: Lazy<PingType> =
-    Lazy::new(|| PingType::new("prototype", true, true, true, true, true, vec![], vec![]));
+pub static PrototypePing: Lazy<PingType> = Lazy::new(|| {
+    PingType::new(
+        "prototype",
+        true,
+        true,
+        true,
+        true,
+        true,
+        vec![],
+        vec![],
+        true,
+    )
+});
 
 fn main() {
     env_logger::init();
@@ -50,6 +61,7 @@ fn main() {
 
     let data_path = PathBuf::from(args.next().expect("need data path"));
 
+    _ = &*PrototypePing;
     let cfg = ConfigurationBuilder::new(true, data_path, "glean.longrunning")
         .with_server_endpoint("invalid-test-host")
         .with_use_core_mps(false)